requirements.txt
Flask
line-bot-sdk
twstock
imgurpython